# backend/specs/openapi.yaml
openapi: 3.1.0
info:
  title: Travel Agent API
  version: 0.1.0
  description: >-
    Authored (spec-first) API contract. The runtime FastAPI schema is asserted to conform to
    this file by tests/test_openapi_contract.py — a route that drifts from this contract turns
    the suite red. Extended one phase at a time; frozen once a phase stabilizes.

paths:
  /api/trips:
    post:
      operationId: create_trip
      summary: Create a trip request (the demo user is resolved server-side, not supplied by the client).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TripRequestCreate' }
      responses:
        '200':
          description: The created trip, status CREATED.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TripRequestOut' }
        '422':
          description: A malformed body, a past depart_date, or a return_date before depart_date.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }

  /api/trips/{trip_id}:
    get:
      operationId: get_trip
      summary: Fetch a single trip by id (the frontend recovers a trip after a page refresh).
      parameters:
        - { name: trip_id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: The trip, in the same shape POST /api/trips returns.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TripRequestOut' }
        '404':
          description: No such trip.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
    patch:
      operationId: update_trip
      summary: Update one or more trip fields. Only fields present in the body are changed.
      parameters:
        - { name: trip_id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TripRequestUpdate' }
      responses:
        '200':
          description: The updated trip.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TripRequestOut' }
        '404':
          description: No such trip.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '422':
          description: >-
            A malformed body, or the resulting depart_date/return_date combination is invalid
            (a past depart_date, or a return_date before depart_date).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }

  /api/trips/{trip_id}/flights/search:
    post:
      operationId: search_trip_flights
      summary: >-
        Search real flight offers for the trip's own stored route and dates. Reuses a real
        search made for the same route+dates within FLIGHT_CACHE_TTL_MINUTES instead of spending
        another unit of the scarce SearchApi quota. Each invocation records an agent run with
        provider, latency, token, cost, and tool-step metrics.
      parameters:
        - { name: trip_id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: >-
            The flight offers found for this trip (possibly empty, with unavailable_reason set —
            the adapter never fabricates an offer).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FlightSearchOut' }
        '404':
          description: No such trip.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '429':
          description: Per-IP rate limit exceeded on this route.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }

  /api/trips/{trip_id}/plan:
    post:
      operationId: plan_trip
      summary: >-
        Generate (or return the already-generated) itinerary. The prompt is built entirely from
        the trip's own stored fields — the client cannot inject free-text prompt content.
      parameters:
        - { name: trip_id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Either the ready itinerary or the agent's clarifying questions.
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: '#/components/schemas/PlanReadyOut' }
                  - { $ref: '#/components/schemas/PlanNeedsClarificationOut' }
                discriminator:
                  propertyName: status
                  mapping:
                    ready: '#/components/schemas/PlanReadyOut'
                    needs_clarification: '#/components/schemas/PlanNeedsClarificationOut'
        '404':
          description: No such trip.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '429':
          description: >-
            Per-IP rate limit exceeded, or the global concurrent-agent-run cap is full (both
            protect scarce LLM quota).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }

  /api/trips/{trip_id}/execution:
    get:
      operationId: get_trip_execution
      summary: >-
        Every flight-search and itinerary-planning agent run for the trip, newest first, with
        its own model/tool steps, API/protocol events, tokens, timing, and cost. The top-level
        event stream is retained for live activity only; execution history groups each event
        under the run that produced it.
      parameters:
        - { name: trip_id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: The execution panel data for this trip.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExecutionPanelOut' }
        '404':
          description: No such trip.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }

  /api/trips/{trip_id}/booking/request:
    post:
      operationId: request_booking
      summary: >-
        Open a HITL booking request for a chosen flight offer (idempotent while active; a
        cancelled request can be opened again).
      parameters:
        - { name: trip_id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BookingRequestCreate' }
      responses:
        '200':
          description: A new or existing active booking log in state PENDING_USER_CONFIRMATION.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookingLogOut' }
        '404':
          description: The trip or the flight offer does not exist.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '422':
          description: The request body failed validation (e.g. wrong field type).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }

  /api/bookings/{log_id}:
    get:
      operationId: get_booking
      summary: The current booking state plus its append-only transition history (what the UI polls).
      parameters:
        - { name: log_id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Current state and transition history.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookingLogOut' }
        '404':
          description: No such booking log.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }

  /api/bookings/{log_id}/confirm:
    post:
      operationId: confirm_booking
      summary: Capture the human's confirmation intent (PENDING -> CONFIRMED). Idempotent replay.
      parameters:
        - { name: log_id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: The booking log, now CONFIRMED (or already CONFIRMED on replay).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookingLogOut' }
        '404':
          description: No such booking log.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '409':
          description: The fare's price-hold TTL has passed, or the transition is illegal.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }

  /api/bookings/{log_id}/execute:
    post:
      operationId: execute_booking
      summary: >-
        The gated write (CONFIRMED -> EXECUTED): store a record locator, plus real booking
        options if the upstream (SearchApi or its recorded cassette) is reachable — a failure to
        fetch those supplementary links degrades honestly (none stored) rather than blocking the
        already-real, human-confirmed execute. Atomic (SELECT ... FOR UPDATE) and idempotent once
        EXECUTED.
      parameters:
        - { name: log_id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: >-
            The booking log, now EXECUTED, with booking_options populated if the upstream link
            fetch succeeded, or empty if it didn't.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookingLogOut' }
        '404':
          description: No such booking log.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '409':
          description: Not in CONFIRMED state, or the fare's price-hold TTL has passed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }

  /api/bookings/{log_id}/cancel:
    post:
      operationId: cancel_booking
      summary: Cancel from PENDING_USER_CONFIRMATION or CONFIRMED (audited transition).
      parameters:
        - { name: log_id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: The booking log, now CANCELLED.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookingLogOut' }
        '404':
          description: No such booking log.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '409':
          description: The booking is already in a terminal state and cannot be cancelled.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }

components:
  schemas:
    BookingState:
      type: string
      enum: [PENDING_USER_CONFIRMATION, CONFIRMED, EXECUTED, CANCELLED, EXPIRED]

    ErrorCode:
      type: string
      description: Machine-readable failure code; the human explanation is in `detail`.
      enum: [booking_not_found, trip_not_found, flight_not_found, booking_expired, invalid_transition, validation_error, rate_limit_exceeded]

    BookingRequestCreate:
      type: object
      required: [flight_search_result_id]
      properties:
        flight_search_result_id: { type: integer }

    BookingTransitionOut:
      type: object
      required: [from_state, to_state, reason, created_at]
      properties:
        from_state: { $ref: '#/components/schemas/BookingState' }
        to_state: { $ref: '#/components/schemas/BookingState' }
        reason: { type: string }
        actor_user_id: { type: [integer, 'null'] }
        created_at: { type: string, format: date-time }

    BookingLogOut:
      type: object
      required: [id, trip_request_id, flight_search_result_id, state, expires_at, created_at]
      properties:
        id: { type: integer }
        trip_request_id: { type: integer }
        flight_search_result_id: { type: integer }
        state: { $ref: '#/components/schemas/BookingState' }
        booking_reference: { type: [string, 'null'] }
        booking_options:
          type: [array, 'null']
          items: { type: object, additionalProperties: true }
        expires_at: { type: string, format: date-time }
        confirmed_at: { type: [string, 'null'], format: date-time }
        executed_at: { type: [string, 'null'], format: date-time }
        created_at: { type: string, format: date-time }
        transitions:
          type: array
          items: { $ref: '#/components/schemas/BookingTransitionOut' }

    ProblemDetail:
      type: object
      required: [code, detail]
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        detail:
          type: string
          description: A human-readable message naming the specific failed precondition and the fix.

    FitnessLevel:
      type: string
      enum: [low, moderate, high]

    TripStatus:
      type: string
      enum: [created, flights_searched, itinerary_ready]

    TripRequestCreate:
      type: object
      required: [origin, destination, destination_airport, depart_date, age, fitness_level]
      properties:
        origin: { type: string, description: 'IATA code, e.g. JFK' }
        destination: { type: string, description: Free-text place name used for activity research }
        destination_airport: { type: string, description: 'IATA code for the flight search' }
        depart_date: { type: string, format: date }
        return_date: { type: [string, 'null'], format: date }
        age: { type: integer }
        fitness_level: { $ref: '#/components/schemas/FitnessLevel' }
        budget_usd: { type: [number, 'null'] }

    TripRequestUpdate:
      type: object
      properties:
        origin: { type: [string, 'null'] }
        destination: { type: [string, 'null'] }
        destination_airport: { type: [string, 'null'] }
        depart_date: { type: [string, 'null'], format: date }
        return_date: { type: [string, 'null'], format: date }
        age: { type: [integer, 'null'] }
        fitness_level:
          oneOf:
            - { $ref: '#/components/schemas/FitnessLevel' }
            - type: 'null'
        budget_usd: { type: [number, 'null'] }

    TripRequestOut:
      type: object
      required: [id, user_id, origin, destination, destination_airport, depart_date, status, created_at]
      properties:
        id: { type: integer }
        user_id: { type: integer }
        origin: { type: string }
        destination: { type: string }
        destination_airport: { type: string }
        depart_date: { type: string, format: date }
        return_date: { type: [string, 'null'], format: date }
        age: { type: [integer, 'null'] }
        fitness_level:
          oneOf:
            - { $ref: '#/components/schemas/FitnessLevel' }
            - type: 'null'
        budget_usd: { type: [number, 'null'] }
        status: { $ref: '#/components/schemas/TripStatus' }
        created_at: { type: string, format: date-time }

    FlightLegOut:
      type: object
      required: [airline, departure_airport, depart_at, arrival_airport, arrive_at]
      properties:
        airline: { type: string }
        flight_number: { type: [string, 'null'] }
        departure_airport: { type: string }
        depart_at: { type: string }
        arrival_airport: { type: string }
        arrive_at: { type: string }
        duration_minutes: { type: [integer, 'null'] }

    FlightOfferOut:
      type: object
      required: [id, offer_index, carrier, price_usd, currency, depart_at, arrive_at, stops, source, legs]
      properties:
        id: { type: integer }
        offer_index: { type: integer }
        carrier: { type: string }
        price_usd: { type: number }
        currency: { type: string }
        depart_at: { type: string }
        arrive_at: { type: string }
        stops: { type: integer }
        source: { type: string, enum: [live, cached] }
        legs:
          description: >-
            Flown segments in travel order. Round-trip offers contain the outbound segments
            followed by the exact selected return segments.
          type: array
          items: { $ref: '#/components/schemas/FlightLegOut' }

    FlightSearchOut:
      type: object
      required: [offers]
      properties:
        offers:
          type: array
          items: { $ref: '#/components/schemas/FlightOfferOut' }
        unavailable_reason: { type: [string, 'null'] }

    ActivityOut:
      type: object
      required: [name, description, intensity, source_url]
      properties:
        name: { type: string }
        description: { type: string }
        intensity: { type: string }
        source_url: { type: string }

    ItineraryDayOut:
      type: object
      required: [day_number, summary, activities]
      properties:
        day_number: { type: integer }
        summary: { type: string }
        activities:
          type: array
          items: { $ref: '#/components/schemas/ActivityOut' }

    ItineraryOut:
      type: object
      required: [days]
      properties:
        days:
          type: array
          items: { $ref: '#/components/schemas/ItineraryDayOut' }

    PlanReadyOut:
      type: object
      required: [status, itinerary]
      properties:
        status: { type: string, enum: [ready] }
        itinerary: { $ref: '#/components/schemas/ItineraryOut' }

    PlanNeedsClarificationOut:
      type: object
      required: [status, questions]
      properties:
        status: { type: string, enum: [needs_clarification] }
        questions:
          type: array
          items: { type: string }

    AgentStepKind:
      type: string
      enum: [model, tool]

    ExecutionEventKind:
      type: string
      enum: [api_call, db_query, protocol, hitl]

    AgentRunStepOut:
      type: object
      required: [seq, kind, name, status]
      properties:
        seq: { type: integer }
        kind: { $ref: '#/components/schemas/AgentStepKind' }
        name: { type: string }
        status: { type: string }
        duration_ms: { type: [integer, 'null'] }
        input_summary: { type: [string, 'null'] }
        output_summary: { type: [string, 'null'] }
        tokens: { type: [integer, 'null'] }

    AgentRunOut:
      type: object
      required: [id, status, model, total_input_tokens, total_output_tokens, total_ms, started_at]
      properties:
        id: { type: integer }
        status: { type: string }
        model: { type: string }
        total_input_tokens: { type: integer }
        total_output_tokens: { type: integer }
        total_ms: { type: integer }
        started_at: { type: string, format: date-time }
        finished_at: { type: [string, 'null'], format: date-time }
        steps:
          type: array
          items: { $ref: '#/components/schemas/AgentRunStepOut' }
        events:
          type: array
          items: { $ref: '#/components/schemas/ExecutionEventOut' }
        estimated_cost_usd: { type: [number, 'null'] }
        budget_utilization_pct: { type: [number, 'null'] }

    ExecutionEventOut:
      type: object
      required: [seq, kind, name, status, detail, created_at]
      properties:
        seq: { type: integer }
        kind: { $ref: '#/components/schemas/ExecutionEventKind' }
        name: { type: string }
        provider: { type: [string, 'null'] }
        status: { type: string }
        detail: { type: string }
        duration_ms: { type: [integer, 'null'] }
        data: { type: [object, 'null'] }
        created_at: { type: string, format: date-time }

    ExecutionPanelOut:
      type: object
      required: [trip_request_id]
      properties:
        trip_request_id: { type: integer }
        agent_runs:
          type: array
          items: { $ref: '#/components/schemas/AgentRunOut' }
        events:
          type: array
          items: { $ref: '#/components/schemas/ExecutionEventOut' }